feat: move text preset, style, size to global maps bb-344
