--- apps/mobile/src/libs/components/text/libs/maps/maps.ts
@@ -1,3 +1,0 @@
-export { fontWeightToFamilyMap } from "./font-weight-to-family-map";
-export { presetToStyleMap } from "./preset-to-style-map";
-export { sizeToStyleMap } from "./size-to-style-map";
