+++ apps/mobile/src/libs/components/text/text.tsx
@@ -3,24 +3,23 @@ import { Text as RNText } from "react-native";
 
 import { BaseColor } from "~/libs/enums/enums";
 import {
+	fontWeightToFamily,
+	presetToTextStyle,
+	sizeToTextStyle,
+} from "~/libs/maps/maps";
+import {
 	type StyleProp,
 	type TextStyle,
 	type ValueOf,
 } from "~/libs/types/types";
 
-import {
-	fontWeightToFamilyMap,
-	presetToStyleMap,
-	sizeToStyleMap,
-} from "./libs/maps/maps";
-
 type Properties = {
 	children: React.ReactNode;
 	color?: ValueOf<typeof BaseColor>;
-	preset?: keyof typeof presetToStyleMap;
-	size?: keyof typeof sizeToStyleMap;
+	preset?: keyof typeof presetToTextStyle;
+	size?: keyof typeof sizeToTextStyle;
 	style?: StyleProp<TextStyle>;
-	weight?: keyof typeof fontWeightToFamilyMap;
+	weight?: keyof typeof fontWeightToFamily;
 };
 
 const Text: React.FC<Properties> = ({
@@ -32,9 +31,9 @@ const Text: React.FC<Properties> = ({
 	weight,
 }: Properties) => {
 	const styles: StyleProp<TextStyle> = [
-		presetToStyleMap[preset],
-		size && sizeToStyleMap[size],
-		weight && fontWeightToFamilyMap[weight],
+		presetToTextStyle[preset],
+		size && sizeToTextStyle[size],
+		weight && fontWeightToFamily[weight],
 		{ color },
 		styleOverride,
 	];
