+++ apps/mobile/src/libs/maps/maps.ts
@@ -1,4 +1,7 @@
 export { categoryToColors } from "./category-to-colors/category-to-colors.map";
 export { colorToGradientColors } from "./color-to-gradient-colors/color-to-gradient-colors.map";
 export { directionToGradient } from "./direction-to-gradient/direction-to-gradient.map";
+export { fontWeightToFamily } from "./font-weight-to-family/font-weight-to-family-map";
+export { presetToTextStyle } from "./preset-to-text-style/preset-to-text-style-map";
 export { sizeToStyles } from "./size-to-styles/size-to-styles";
+export { sizeToTextStyle } from "./size-to-text-style/size-to-text-style.map";
